+++ packages/snjs/lib/index.ts
@@ -25,11 +25,13 @@ export { KeyRecoveryStrings, SessionStrings } from './services/api/messages';
 export type { RemoteSession } from './services/api/session';
 
 export { SNApplication } from '@Lib/application';
+
 export { SNProtocolService, KeyMode } from '@Services/protocol_service';
 export { SNProtocolOperator001 } from '@Protocol/operator/001/operator_001';
 export { SNProtocolOperator002 } from '@Protocol/operator/002/operator_002';
 export { SNProtocolOperator003 } from '@Protocol/operator/003/operator_003';
 export { SNProtocolOperator004 } from '@Protocol/operator/004/operator_004';
+
 export { SNRootKey } from '@Protocol/root_key';
 export { SNRootKeyParams } from './protocol/key_params';
 export { DeviceInterface } from '@Lib/device_interface';
@@ -169,7 +171,7 @@ export {
   dateToLocalizedString,
   nonSecureRandomIdentifier,
   sanitizeHtmlString,
-  lastElement
+  lastElement,
 } from './utils';
 export { Uuid } from '@Lib/uuid';
 export {
@@ -211,6 +213,7 @@ export {
   PayloadByMerging,
   CopyPayload,
 } from '@Payloads/generator';
+export { PayloadsByAlternatingUuid } from './protocol/payloads';
 export {
   PayloadSource,
   isPayloadSourceRetrieved,
@@ -258,3 +261,5 @@ export {
 export { Migration2_20_0 } from './migrations/2_20_0';
 export { Migration2_42_0 } from './migrations/2_42_0';
 export { ApplicationStage } from '@Lib/stages';
+
+export { ActionVerb } from '@Lib/models/app/extension';