+++ packages/snjs/mocha/lib/factory.js
@@ -18,8 +18,8 @@ export async function createAndInitSimpleAppContext(
   }
 ) {
   const application = await createInitAppWithRandNamespace(environment);
-  const email = Uuid.GenerateUuidSynchronously();
-  const password = Uuid.GenerateUuidSynchronously();
+  const email = Uuid.GenerateUuid();
+  const password = Uuid.GenerateUuid();
   const newPassword = randomString();
 
   if (registerUser) {
@@ -43,8 +43,8 @@ export async function createAppContext(identifier) {
     identifier = `${Math.random()}`;
   }
   const application = await createApplication(identifier);
-  const email = Uuid.GenerateUuidSynchronously();
-  const password = Uuid.GenerateUuidSynchronously();
+  const email = Uuid.GenerateUuid();
+  const password = Uuid.GenerateUuid();
   const passcode = 'mypasscode';
   const handleChallenge = (challenge) => {
     const responses = [];
@@ -403,7 +403,7 @@ export function createItemParams(contentType) {
 
 export function generateUuid() {
   const crypto = new SNWebCrypto();
-  return crypto.generateUUIDSync();
+  return crypto.generateUUID();
 }
 
 export function createNoteParams({ title, text, dirty = true } = {}) {
@@ -601,4 +601,19 @@ export async function pinNote(application, note) {
   return application.changeItem(note.uuid, (mutator) => {
     mutator.pinned = true;
   });
+}
+
+export async function alternateUuidForItem(application, uuid) {
+  const item = application.itemManager.findItem(uuid);
+  const payload = CreateMaxPayloadFromAnyObject(item);
+  const results = await PayloadsByAlternatingUuid(
+    payload,
+    application.payloadManager.getMasterCollection()
+  );
+  await application.payloadManager.emitPayloads(
+    results,
+    PayloadSource.LocalChanged
+  );
+  await application.syncService.persistPayloads(results);
+  return application.itemManager.findItem(results[0].uuid);
 }