+++ packages/snjs/mocha/predicate.test.js
@@ -36,7 +36,7 @@ const createItemParams = function (title = 'Hello', desc = 'World') {
 describe('predicates', async function () {
   before(async function () {
     const crypto = new SNWebCrypto();
-    Uuid.SetGenerators(crypto.generateUUIDSync, crypto.generateUUID);
+    Uuid.SetGenerators(crypto.generateUUID, crypto.generateUUID);
   });
 
   beforeEach(async function () {