+++ packages/snjs/mocha/singletons.test.js
@@ -15,7 +15,7 @@ describe('singletons', function () {
 
   function createPrefsPayload() {
     const params = {
-      uuid: Uuid.GenerateUuidSynchronously(),
+      uuid: Uuid.GenerateUuid(),
       content_type: ContentType.UserPrefs,
       content: {
         foo: 'bar',
@@ -36,8 +36,8 @@ describe('singletons', function () {
     localStorage.clear();
     this.expectedItemCount = BASE_ITEM_COUNT;
     this.application = await Factory.createInitAppWithRandNamespace();
-    this.email = Uuid.GenerateUuidSynchronously();
-    this.password = Uuid.GenerateUuidSynchronously();
+    this.email = Uuid.GenerateUuid();
+    this.password = Uuid.GenerateUuid();
     this.registerUser = async () => {
       await Factory.registerUserToApplication({
         application: this.application,
@@ -163,8 +163,8 @@ describe('singletons', function () {
   it('resolves registered predicate with signing in/out', async function () {
     await this.registerUser();
     await this.signOut();
-    this.email = Uuid.GenerateUuidSynchronously();
-    this.password = Uuid.GenerateUuidSynchronously();
+    this.email = Uuid.GenerateUuid();
+    this.password = Uuid.GenerateUuid();
     await this.createExtMgr();
     this.expectedItemCount += 1;
     await Factory.registerUserToApplication({
@@ -354,7 +354,7 @@ describe('singletons', function () {
       payload.content
     );
     const originalUuid = resolvedItem.uuid;
-    await this.application.syncService.alternateUuidForItem(resolvedItem.uuid);
+    await Factory.alternateUuidForItem(this.application, resolvedItem.uuid);
     await this.application.syncService.sync(syncOptions);
     const resolvedItem2 = await this.application.singletonManager.findOrCreateSingleton(
       predicate,