+++ packages/components/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/standardnotes/snjs/compare/@standardnotes/components@1.7.2...@standardnotes/components@1.7.3) (2022-02-16)
+
+**Note:** Version bump only for package @standardnotes/components
+
+
+
+
+
+## [1.7.2](https://github.com/standardnotes/snjs/compare/@standardnotes/components@1.7.1...@standardnotes/components@1.7.2) (2022-02-15)
+
+**Note:** Version bump only for package @standardnotes/components
+
+
+
+
+
+## [1.7.1](https://github.com/standardnotes/snjs/compare/@standardnotes/components@1.7.0...@standardnotes/components@1.7.1) (2022-02-14)
+
+**Note:** Version bump only for package @standardnotes/components
+
+
+
+
+
 # [1.7.0](https://github.com/standardnotes/snjs/compare/@standardnotes/components@1.6.0...@standardnotes/components@1.7.0) (2022-02-08)